+++ packages/codegen/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## V3
+
+### Changed
+
+- Extensions now include all extensions available in the chain.
+- Descriptors now include "non-default extensions"
+
 ## Unreleased
 
 ## 0.22.3 to 0.22.5 - 2026-05-19